Вынесена логика создания алерта для сортировки
