+++ FakeNFT/Modules/Cart/CartViewRouter.swift
@@ -41,36 +41,7 @@ extension CartViewRouter: CartViewRouterProtocol {
         viewController: UIViewController,
         onChoosingSortingTrait: @escaping (CartOrderSorter.SortingTrait) -> Void
     ) {
-        let alertController = UIAlertController(
-            title: "CART_SORT_ALERT_TITLE".localized,
-            message: nil,
-            preferredStyle: .actionSheet
-        )
-
-        let sortByPriceAction = UIAlertAction(
-            title: "CART_SORT_ALERT_PRICE_ACTION_TITLE".localized,
-            style: .default
-        ) { _ in onChoosingSortingTrait(.price) }
-
-        let sortByRatingAction = UIAlertAction(
-            title: "CART_SORT_ALERT_RATING_ACTION_TITLE".localized,
-            style: .default
-        ) { _ in onChoosingSortingTrait(.rating) }
-
-        let sortByNameAction = UIAlertAction(
-            title: "CART_SORT_ALERT_NAME_ACTION_TITLE".localized,
-            style: .default
-        ) { _ in onChoosingSortingTrait(.name) }
-
-        let closeAction = UIAlertAction(
-            title: "CART_SORT_ALERT_CLOSE_ACTION_TITLE".localized,
-            style: .cancel
-        ) { _ in }
-
-        [sortByPriceAction, sortByRatingAction, sortByNameAction, closeAction].forEach {
-            alertController.addAction($0)
-        }
-
+        let alertController = UIAlertController.sortingAlertController(onChoosingSortingTrait: onChoosingSortingTrait)
         viewController.present(alertController, animated: true)
     }
